Ask user to confirm deleting

+++ src/Command/Task/DeleteCommand.php
@@ -6,6 +6,7 @@ use AlVi\Command\AbstractCommand;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\Console\Question\ConfirmationQuestion;
 
 class DeleteCommand extends AbstractCommand
 {
@@ -26,7 +27,12 @@ class DeleteCommand extends AbstractCommand
         $taskList = $this->resolveTaskList($input, $output);
         $task = $this->resolveTask($taskList, $input, $output);
 
-        // @TODO Add confirmation before actual deleting
+        $helper = $this->getHelper('question');
+        $question = new ConfirmationQuestion('Are you sure you want to delete task? Y/N: ', false);
+        if (!$helper->ask($input, $output, $question)) {
+            return;
+        }
+
         $service = $this->getTasksGoogleService();
         $service->tasks->delete($taskList->getId(), $task->getId());
